requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docutils==0.17.1
idna==2.10
iniconfig==1.1.1
lockfile==0.12.2
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==0.17.0
regex==2021.4.4
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
